[Silver II] Title: 가장 긴 증가하는 부분 수열, Time: 124 ms, Memory: 14580 KB -BaekjoonHub

+++ 백준/Silver/11053. 가장 긴 증가하는 부분 수열/가장 긴 증가하는 부분 수열.java
@@ -21,27 +21,27 @@ public class Main {
     }
 
     public static void mySolution() throws IOException {
-        final int N = Integer.parseInt(br.readLine());
-        int[] numbers = new int[N];
+        final short N = Short.parseShort(br.readLine());
+        short[] numbers = new short[N];
         StringTokenizer stringTokenizer = new StringTokenizer(br.readLine(), " ");
 
-        for (int i = 0; i < N; ++i) {
-            numbers[i] = Integer.parseInt(stringTokenizer.nextToken());
+        for (short i = 0; i < N; ++i) {
+            numbers[i] = Short.parseShort(stringTokenizer.nextToken());
         }
 
         // numbers[i]를 마지막 원소로 하는 가장 긴 증가 부분 수열의 길이
-        int[] dp = new int[N];
-        Arrays.fill(dp, 1); // 자기 자신만 포함된 부분 배열도 성립. 1로 초기화
-        int maxLength = 1;
+        short[] dp = new short[N];
+        Arrays.fill(dp, (short) 1);  // 자기 자신만 포함된 부분 배열도 성립. 1로 초기화
+        short maxLength = 1;
 
-        for (int i = 0; i < N; ++i) {
-            for (int j = 0; j < i; ++j) {
+        for (short i = 0; i < N; ++i) {
+            for (short j = 0; j < i; ++j) {
                 if (numbers[i] > numbers[j]) {
-                    dp[i] = Math.max(dp[i], dp[j] + 1);
+                    dp[i] = (short) Math.max(dp[i], dp[j] + 1);
                 }
             }
 
-            maxLength = Math.max(maxLength, dp[i]);
+            maxLength = (short) Math.max(maxLength, dp[i]);
         }
 
         bw.write(String.valueOf(maxLength));
